scripts/hdxloader/requirements.txt
hdx-python-api>=5.7.2
python-slugify>=6.1.2